+++ src/Schema/Postgres/Types/XmlType.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Php\Support\Laravel\Database\Schema\Postgres\Types;
+
+use Doctrine\DBAL\Platforms\AbstractPlatform;
+use Doctrine\DBAL\Types\Type;
+
+/**
+ * Class XmlType
+ * @package Php\Support\Laravel
+ */
+class XmlType extends Type
+{
+    public const TYPE_NAME = 'xml';
+
+    public function getSQLDeclaration(array $fieldDeclaration, AbstractPlatform $platform)
+    {
+        return static::TYPE_NAME;
+    }
+
+    public function getName(): string
+    {
+        return self::TYPE_NAME;
+    }
+}